+++ src/shader_renderer/mod.rs
@@ -1,0 +1,43 @@
+//! Shader rendering system for shekere
+//!
+//! This module provides Material2d-based rendering with dynamic shader loading,
+//! multi-pass rendering, and persistent (trail) effects.
+
+mod generation;
+mod materials;
+mod setup;
+mod state;
+
+use bevy::prelude::*;
+use bevy::sprite::Material2dPlugin;
+use materials::{ShekereShaderMaterial, ShekereShaderMaterialPass0, ShekereShaderMaterialPass1};
+use setup::setup_dynamic_shader_system;
+use state::{MultiPassState, PersistentPassState};
+
+// Re-export for external use
+pub use generation::check_shader_reload;
+pub use setup::{update_multipass_uniforms, update_persistent_uniforms, update_shader_uniforms};
+pub use state::InputBufferHandles;
+
+/// Plugin for shader rendering with dynamic loading and multi-pass support
+pub struct SimpleShaderRenderPlugin;
+
+impl Plugin for SimpleShaderRenderPlugin {
+    fn build(&self, app: &mut App) {
+        app.add_plugins((
+            Material2dPlugin::<ShekereShaderMaterial>::default(),
+            Material2dPlugin::<ShekereShaderMaterialPass0>::default(),
+            Material2dPlugin::<ShekereShaderMaterialPass1>::default(),
+        ))
+        .add_systems(Startup, setup_dynamic_shader_system)
+        .add_systems(
+            Update,
+            (
+                update_shader_uniforms,
+                update_multipass_uniforms.run_if(resource_exists::<MultiPassState>),
+                update_persistent_uniforms.run_if(resource_exists::<PersistentPassState>),
+                check_shader_reload,
+            ),
+        );
+    }
+}
